CC/Deploy model ML/doctor/requirements.txt
Flask==2.1.0
gunicorn==20.1.0
numpy==1.22.4
Pillow==9.1.1
opencv-python==4.5.5.64
tensorflow